+++ core/services/identity-resolver.service.test.ts
@@ -1,27 +1,17 @@
 /**
- * Identity Service Tests
+ * Identity Resolver Service Tests
  *
- * Tests for the Identity Service (Developer Identity Resolution).
+ * Tests for the Identity Resolver Service (Developer Identity Resolution).
  * These tests verify that:
  * - resolveDeveloperByAccount() works correctly (PRIMARY method)
  * - resolveDeveloperByIdentifier() works correctly (SECONDARY method)
  * - findDuplicates() calculates confidence scores correctly
- * - mergeDevelopers() handles transactions and audit logs correctly
- * - addIdentifier() handles duplicates correctly
- * - removeIdentifier() deletes identifiers correctly
  *
  * Test Strategy:
  * - NO MOCKS: Tests use real database connection
  * - Cleanup: Tests clean up their own data
  * - Verification: Actual database records are checked
  * - RLS: Tests verify tenant isolation
- * - Transaction testing: Verifies rollback on errors
- *
- * Setup:
- * - Requires PostgreSQL running (docker-compose)
- * - Uses DATABASE_* environment variables
- * - Sets tenant context before all tests
- * - Closes connections after tests
  */
 
 import { describe, it, expect, beforeAll, afterAll } from 'vitest';
@@ -32,18 +22,15 @@ import {
   clearTenantContext,
 } from '../db/connection.js';
 import * as schema from '../db/schema/index.js';
-import { eq } from 'drizzle-orm';
 import {
   resolveDeveloperByAccount,
   resolveDeveloperByIdentifier,
   findDuplicates,
-  mergeDevelopers,
-  addIdentifier,
-  removeIdentifier,
-} from './identity.service.js';
+} from './identity-resolver.service.js';
+import { addIdentifier } from './identity-identifiers.service.js';
 import { createDeveloper, deleteDeveloper } from './drm.service.js';
 
-describe('Identity Service', () => {
+describe('Identity Resolver Service', () => {
   beforeAll(async () => {
     // Set tenant context for all tests
     // This is REQUIRED because RLS is enabled
@@ -389,321 +376,6 @@ describe('Identity Service', () => {
 
       // Cleanup
       await deleteDeveloper('default', dev.developerId);
-    });
-  });
-
-  describe('mergeDevelopers', () => {
-    it('should merge two developers successfully', async () => {
-      // Arrange: Create two developers
-      const timestamp = Date.now();
-      const target = await createDeveloper('default', {
-        displayName: 'Target Developer',
-        primaryEmail: `target-${timestamp}@example.com`,
-        orgId: null,
-        tags: ['target-tag'],
-      });
-      const source = await createDeveloper('default', {
-        displayName: 'Source Developer',
-        primaryEmail: `source-${timestamp}@example.com`,
-        orgId: null,
-        tags: ['source-tag'],
-      });
-
-      // Add identifier to source
-      await addIdentifier('default', source.developerId, {
-        kind: 'email',
-        value: `source-${timestamp}@example.com`,
-      });
-
-      // Act: Merge source into target
-      const result = await mergeDevelopers('default', {
-        intoDeveloperId: target.developerId,
-        fromDeveloperId: source.developerId,
-        reason: 'Test merge',
-      });
-
-      // Assert: Should return target developer
-      expect(result.developerId).toBe(target.developerId);
-
-      // Assert: Tags should be merged
-      expect(result.tags).toContain('target-tag');
-      expect(result.tags).toContain('source-tag');
-
-      // Assert: Source developer should be deleted
-      const db = getDb();
-      const sourceAfterMerge = await db
-        .select()
-        .from(schema.developers)
-        .where(eq(schema.developers.developerId, source.developerId))
-        .limit(1);
-      expect(sourceAfterMerge.length).toBe(0);
-
-      // Assert: Identifiers should be moved to target
-      const targetIdentifiers = await db
-        .select()
-        .from(schema.developerIdentifiers)
-        .where(eq(schema.developerIdentifiers.developerId, target.developerId));
-      expect(targetIdentifiers.length).toBeGreaterThan(0);
-
-      // Cleanup
-      await deleteDeveloper('default', target.developerId);
-    });
-
-    // Note: merge log creation is tested indirectly through the "should merge two developers successfully" test
-    // Direct testing of merge logs has issues with RLS and transaction contexts, so we skip it here.
-
-    it('should throw error if target developer not found', async () => {
-      // Arrange: Create only source developer
-      const timestamp = Date.now();
-      const source = await createDeveloper('default', {
-        displayName: 'Source Only',
-        primaryEmail: `source-only-${timestamp}@example.com`,
-        orgId: null,
-      });
-
-      // Act & Assert: Merge with non-existent target should fail
-      await expect(
-        mergeDevelopers('default', {
-          intoDeveloperId: '99999999-9999-4999-8999-999999999999',
-          fromDeveloperId: source.developerId,
-          reason: 'Should fail',
-        })
-      ).rejects.toThrow(/target.*not found/i);
-
-      // Cleanup
-      await deleteDeveloper('default', source.developerId);
-    });
-
-    it('should throw error if source developer not found', async () => {
-      // Arrange: Create only target developer
-      const timestamp = Date.now();
-      const target = await createDeveloper('default', {
-        displayName: 'Target Only',
-        primaryEmail: `target-only-${timestamp}@example.com`,
-        orgId: null,
-      });
-
-      // Act & Assert: Merge with non-existent source should fail
-      await expect(
-        mergeDevelopers('default', {
-          intoDeveloperId: target.developerId,
-          fromDeveloperId: '99999999-9999-4999-8999-999999999999',
-          reason: 'Should fail',
-        })
-      ).rejects.toThrow(/source.*not found/i);
-
-      // Cleanup
-      await deleteDeveloper('default', target.developerId);
-    });
-
-    it('should throw error when merging developer with itself', async () => {
-      // Arrange: Create developer
-      const timestamp = Date.now();
-      const dev = await createDeveloper('default', {
-        displayName: 'Self Merge Test',
-        primaryEmail: `self-merge-${timestamp}@example.com`,
-        orgId: null,
-      });
-
-      // Act & Assert: Merge with itself should fail
-      await expect(
-        mergeDevelopers('default', {
-          intoDeveloperId: dev.developerId,
-          fromDeveloperId: dev.developerId,
-          reason: 'Should fail',
-        })
-      ).rejects.toThrow(/cannot merge developer with itself/i);
-
-      // Cleanup
-      await deleteDeveloper('default', dev.developerId);
-    });
-  });
-
-  describe('addIdentifier', () => {
-    it('should add identifier to developer', async () => {
-      // Arrange: Create developer
-      const timestamp = Date.now();
-      const dev = await createDeveloper('default', {
-        displayName: 'Add Identifier Test',
-        primaryEmail: `add-id-${timestamp}@example.com`,
-        orgId: null,
-      });
-
-      // Act: Add identifier
-      const identifier = await addIdentifier('default', dev.developerId, {
-        kind: 'email',
-        value: `add-id-${timestamp}@example.com`,
-      }, 1.0);
-
-      // Assert: Identifier should be created
-      expect(identifier.developerId).toBe(dev.developerId);
-      expect(identifier.kind).toBe('email');
-      expect(identifier.valueNormalized).toBe(`add-id-${timestamp}@example.com`);
-      expect(Number(identifier.confidence)).toBe(1.0);
-
-      // Cleanup
-      await deleteDeveloper('default', dev.developerId);
-    });
-
-    it('should normalize identifier value', async () => {
-      // Arrange: Create developer
-      const timestamp = Date.now();
-      const dev = await createDeveloper('default', {
-        displayName: 'Normalize Test',
-        primaryEmail: null,
-        orgId: null,
-      });
-
-      // Act: Add identifier with uppercase email
-      const identifier = await addIdentifier('default', dev.developerId, {
-        kind: 'email',
-        value: `UPPERCASE-${timestamp}@EXAMPLE.COM`,
-      });
-
-      // Assert: Should be normalized to lowercase
-      expect(identifier.valueNormalized).toBe(`uppercase-${timestamp}@example.com`);
-
-      // Cleanup
-      await deleteDeveloper('default', dev.developerId);
-    });
-
-    it('should update existing identifier if same developer', async () => {
-      // Arrange: Create developer with identifier
-      const timestamp = Date.now();
-      const dev = await createDeveloper('default', {
-        displayName: 'Update Identifier Test',
-        primaryEmail: null,
-        orgId: null,
-      });
-
-      const email = `update-id-${timestamp}@example.com`;
-      await addIdentifier('default', dev.developerId, {
-        kind: 'email',
-        value: email,
-      }, 0.8);
-
-      // Act: Add same identifier with different confidence
-      const updated = await addIdentifier('default', dev.developerId, {
-        kind: 'email',
-        value: email,
-      }, 0.9);
-
-      // Assert: Should update confidence
-      expect(Number(updated.confidence)).toBe(0.9);
-      expect(updated.developerId).toBe(dev.developerId);
-
-      // Cleanup
-      await deleteDeveloper('default', dev.developerId);
-    });
-
-    it('should throw error on duplicate identifier (different developer)', async () => {
-      // Arrange: Create two developers
-      const timestamp = Date.now();
-      const dev1 = await createDeveloper('default', {
-        displayName: 'Dev 1',
-        primaryEmail: null,
-        orgId: null,
-      });
-      const dev2 = await createDeveloper('default', {
-        displayName: 'Dev 2',
-        primaryEmail: null,
-        orgId: null,
-      });
-
-      const sharedEmail = `shared-conflict-${timestamp}@example.com`;
-
-      // Add identifier to dev1
-      await addIdentifier('default', dev1.developerId, {
-        kind: 'email',
-        value: sharedEmail,
-      });
-
-      // Act & Assert: Adding same identifier to dev2 should fail
-      await expect(
-        addIdentifier('default', dev2.developerId, {
-          kind: 'email',
-          value: sharedEmail,
-        })
-      ).rejects.toThrow(/conflict/i);
-
-      // Cleanup
-      await deleteDeveloper('default', dev1.developerId);
-      await deleteDeveloper('default', dev2.developerId);
-    });
-
-    it('should throw error for invalid confidence score', async () => {
-      // Arrange: Create developer
-      const timestamp = Date.now();
-      const dev = await createDeveloper('default', {
-        displayName: 'Invalid Confidence Test',
-        primaryEmail: null,
-        orgId: null,
-      });
-
-      // Act & Assert: Confidence > 1.0 should fail
-      await expect(
-        addIdentifier('default', dev.developerId, {
-          kind: 'email',
-          value: `test-${timestamp}@example.com`,
-        }, 1.5)
-      ).rejects.toThrow(/confidence must be between/i);
-
-      // Act & Assert: Confidence < 0.0 should fail
-      await expect(
-        addIdentifier('default', dev.developerId, {
-          kind: 'email',
-          value: `test2-${timestamp}@example.com`,
-        }, -0.1)
-      ).rejects.toThrow(/confidence must be between/i);
-
-      // Cleanup
-      await deleteDeveloper('default', dev.developerId);
-    });
-  });
-
-  describe('removeIdentifier', () => {
-    it('should remove identifier from developer', async () => {
-      // Arrange: Create developer with identifier
-      const timestamp = Date.now();
-      const dev = await createDeveloper('default', {
-        displayName: 'Remove Identifier Test',
-        primaryEmail: null,
-        orgId: null,
-      });
-
-      const identifier = await addIdentifier('default', dev.developerId, {
-        kind: 'email',
-        value: `remove-id-${timestamp}@example.com`,
-      });
-
-      // Act: Remove identifier
-      const result = await removeIdentifier('default', identifier.identifierId);
-
-      // Assert: Should return true
-      expect(result).toBe(true);
-
-      // Verify deletion
-      const db = getDb();
-      const deleted = await db
-        .select()
-        .from(schema.developerIdentifiers)
-        .where(eq(schema.developerIdentifiers.identifierId, identifier.identifierId))
-        .limit(1);
-      expect(deleted.length).toBe(0);
-
-      // Cleanup
-      await deleteDeveloper('default', dev.developerId);
-    });
-
-    it('should return false for non-existent identifier', async () => {
-      // Act: Try to remove non-existent identifier
-      const result = await removeIdentifier(
-        'default',
-        '99999999-9999-4999-8999-999999999999'
-      );
-
-      // Assert: Should return false
-      expect(result).toBe(false);
     });
   });
 });
